refactor(runtime): extract session branching into session-branch.ts
Pure move from session-manager.ts (#1084, runtime/other lane):
- createBranchSession, cloneBranchRuntimeLedger, the turn-boundary
  message copy helpers, and the run/event clone helpers now live in
  session-branch.ts behind a minimal SessionBranchDeps seam.
- createBranchSession returns the child SessionHeader; SessionManager
  wraps it with headerToSummary, so the new leaf never imports back
  into session-manager.ts.
- headerLineage is now imported from agent-run-recovery.ts instead of
  keeping a verbatim private copy.

Behavior-neutral; @maka/runtime public exports unchanged.
